+++ common/src/semblance/algorithm/StretchFreeAlgorithm.cpp
@@ -38,9 +38,9 @@ void StretchFreeAlgorithm::computeSemblanceAndParametersForMidpoint(float m0) {
 
     deviceResultArray->pasteTo(computedResults);
 
-    unsigned long totalUsedTracesCount = static_cast<unsigned long>(filteredTracesCount) *
-            static_cast<unsigned long>(numberOfSamplesPerTrace) *
-            static_cast<unsigned long>(totalNumberOfParameters);
+    float totalUsedTracesCount = static_cast<float>(filteredTracesCount) *
+            static_cast<float>(numberOfSamplesPerTrace) *
+            static_cast<float>(totalNumberOfParameters);
 
     saveStatisticalResults(totalUsedTracesCount, totalExecutionTime, selectionExecutionTime);
 }
